refactor(pass-7): add core public type seam
