+++ src/core/index.ts
@@ -1,3 +1,4 @@
 export { LAMBDA, calculateFreshnessScore, scoreLabel } from "./decay.js";
 export { looksLikeFailedAdapterContent } from "./guards.js";
 export { stampFreshness, toStructuredJSON, formatForLLM } from "./envelope.js";
+export type { FreshContext, ExtractOptions, AdapterResult } from "./types.js";
