+++ src/core/types.ts
@@ -1,0 +1,25 @@
+export interface FreshContext {
+  content: string;
+  source_url: string;
+  content_date: string | null;
+  retrieved_at: string;
+  freshness_confidence: "high" | "medium" | "low";
+  freshness_score: number | null;
+  adapter: string;
+}
+
+export interface ExtractOptions {
+  url: string;
+  prompt?: string;
+  maxLength?: number;
+  location?: string;
+  remoteOnly?: boolean;
+  maxAgeDays?: number;
+  keywords?: string[];
+}
+
+export interface AdapterResult {
+  raw: string;
+  content_date: string | null;
+  freshness_confidence: "high" | "medium" | "low";
+}
